requirements.txt
# ITU Racing - FSAE Signal Analysis Tool
# Python Dependencies
# Install with: pip install -r requirements.txt

# GUI Framework
customtkinter>=5.2.0
CTkMessagebox>=2.5

# Plotting & Visualization
plotly>=5.18.0
matplotlib>=3.7.0
pillow>=10.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # For Excel file support

# Signal Processing
scipy>=1.11.0

# Optional: For better performance on large datasets
# numba>=0.58.0
